move profile fetch up to app.tsx
