+++ src/App.tsx
@@ -18,6 +18,7 @@ import ProtectedRoute from "./Components/ProtectedRoute/ProtectedRoute";
 
 // services
 import * as authService from "./Services/authService";
+import * as profileService from '../src/Services/profileService'
 
 // stylesheets
 import "./css/app.css";
@@ -41,6 +42,33 @@ function App(): JSX.Element {
     const handleAuthEvt = (): void => {
         setUser(authService.getUser());
     };
+
+  const profileId = user!.profile.id
+
+  const [profile, setProfile] = useState<Profile>({
+    name: "",
+    userName: "",
+    photo: "",
+    id: 0,
+    createdAt: "",
+    updatedAt: "",
+  })
+
+  useEffect((): void => {
+    const fetchProfile = async (): Promise<void> => {
+      try {
+        console.log("profileId", profileId)
+        const profileData: Profile = await profileService.getProfile(profileId)
+        console.log("profileData", profileData)
+        setProfile(profileData)
+      } catch (error) {
+        console.log(error)
+      }
+    }
+    fetchProfile()
+  }, [user])
+
+
 
     return (
         <>
@@ -69,7 +97,7 @@ function App(): JSX.Element {
                     path="/profile"
                     element={
                         <ProtectedRoute user={user}>
-                            <ProfilePage user={user} />
+                            <ProfilePage user={user} profile={profile}/>
                         </ProtectedRoute>
                     }
                 />
